service RpcGoodsService update goodsquanlityvalue and mapper update5

+++ supermall-common/supermall-common-dao/src/main/java/com/zb/mapper/GoodsMapper.java
@@ -8,7 +8,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface GoodsMapper {
 
-	public Goods getGoodsById(@Param(value = "id") Long id)throws Exception;
+	public Goods getGoodsById(@Param(value = "id") Integer id)throws Exception;
 
 	public List<Goods>	getGoodsListByMap(Map<String,Object> param)throws Exception;
